requirements.txt
dj-database-url==0.5.0
Django==1.11.15
django-crispy-forms==1.7.2
gunicorn==19.9.0
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
whitenoise==4.0
django-ckeditor==5.6.1
django-import-export==1.0.1
